Refactor : 쿠키 네이밍 세션으로 변경
서버에서 사용하는 것이기 때문에 세션으로 이름 변경
