+++ src/main/java/controller/URLController.java
@@ -1,7 +1,7 @@
 package controller;
 
 import config.AppConfig;
-import cookie.Cookie;
+import session.Session;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import request.HttpRequest;
@@ -16,7 +16,7 @@ public class URLController {
     private final UserController userController = AppConfig.userController();
 
 
-    public String mapUrl(String path, HttpRequest httpRequest, HttpResponse httpResponse, BufferedReader br, Cookie cookie) throws IOException {
+    public String mapUrl(String path, HttpRequest httpRequest, HttpResponse httpResponse, BufferedReader br, Session cookie) throws IOException {
         // localhost:8080 기본화면으로 이동
         if (path.equals("/")) {
             httpResponse.setStatus(200);
