+++ src/main/java/webserver/RequestHandler.java
@@ -6,7 +6,7 @@ import java.nio.charset.StandardCharsets;
 
 import config.AppConfig;
 import controller.URLController;
-import cookie.Cookie;
+import session.Session;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import request.HttpRequest;
@@ -34,7 +34,7 @@ public class RequestHandler implements Runnable {
             HttpRequest httpRequest = new HttpRequest();
             httpRequest.setRequestLine(br.readLine());
             HttpResponse httpResponse = new HttpResponse();
-            Cookie cookie = new Cookie();
+            Session cookie = new Session();
 
             String path = httpRequest.getUrl();
             path = urlController.mapUrl(path, httpRequest, httpResponse, br, cookie);
